Removed UserManager and moved functionality into TollBot.
